requirements.txt
streamlit==1.28.0
openai==0.28.1
python-dotenv